requirements.txt
# # Google Gmail API + Auth
# google-api-python-client
# google-auth-httplib2
# google-auth-oauthlib
# google.auth.transport.requests
# google.oauth2.credentials
# google_auth_oauthlib.flow
# googleapiclient.discovery
# LangChain ecosystem
langchain
langchain-community
langchain_google_genai

# Vector store (FAISS)
faiss-cpu==1.8.0

# Environment variables (optional)
python-dotenv
langchain_chroma
sentence-transformers
langchain_huggingface


